renaming the specialized algorithms, bugfixes
