+++ relative_localization/ros/include/relative_localization/checkerboard_localization.h
