+++ relative_localization/ros/launch/checkerboard_localization.launch
@@ -9,8 +9,9 @@
 	<node pkg="tf" type="static_transform_publisher" name="static_transform_publisher_checkerboard_reflector" output="screen" args="0.0 -0.7 0 0 0 0  checkerboard_reference checkerboard 100"/>
 	<node pkg="tf" type="static_transform_publisher" name="static_transform_publisher_checkerboard_reflector_nav" output="screen" args="0.0 0.0 0.0 -1.57079632679 0 -1.57079632679  checkerboard_reference_nav checkerboard_reference 100"/>
 
-	<node name="checkerboard_localisation" pkg="visualization_msgs" type="checkerboard_localisation_node" output="screen">
-		<rosparam command="load" file="$(find visualization_msgs)/ros/launch/checkerboard_localisation_params.yaml"/>
-		<remap from="laser_scan_in" to="/scan"/>
+	<node ns="checkerboard_localization" name="relative_localization" pkg="relative_localization" type="relative_localization_node" output="screen">
+		<rosparam command="load" file="$(find relative_localization)/ros/launch/checkerboard_localization_params.yaml"/>
+		<param name="localization_method" value="checkerboard" />
+		<remap from="laser_scan_in" to="/scan" />
 	</node>
 </launch>
