+++ relative_localization/ros/launch/checkerboard_localization_params.yaml
